Chore: add deposit, withdrawal, read depositlogs

+++ src/main/java/org/example/first/groundingappapis/entity/DepositLog.java
@@ -40,7 +40,7 @@ public class DepositLog {
     private String type;
 
     @Column(name = "amount")
-    private Integer amount;
+    private Long amount;
 
     @CreatedDate
     @Column(name = "created_at", nullable = false)
@@ -52,13 +52,13 @@ public class DepositLog {
     }
 
     //입금
-    public void setLogOfDeposit(Integer amount) {
+    public void setLogOfDeposit(Long amount) {
         this.type = "입금";
         this.amount = amount;
     }
 
     //출금
-    public void setLogOfWithdraw(Integer amount) {
+    public void setLogOfWithdraw(Long amount) {
         this.type = "출금";
         this.amount = amount;
     }
@@ -71,7 +71,7 @@ public class DepositLog {
     }
 
     @Builder
-    public DepositLog(Account account, String type, Integer amount, LocalDateTime createdAt) {
+    public DepositLog(Account account, String type, Long amount, LocalDateTime createdAt) {
         this.account = account;
         this.type = type;
         this.amount = amount;
